% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt4pv-package.R
\docType{package}
\name{adapt4pv-package}
\alias{adapt4pv-package}
\title{Adaptive approaches for signal detection in PharmacoVigilance}
\description{
This package fits adaptive lasso approaches in high dimension for signal detection in
pharmacovigilance.
In addition to classical implementations found in the litterature, we implemented
two approaches particularly appropriated to variable selections framework, which
is the one that stands in pharmacovigilance.
We also supply in this package signal detection approaches based on lasso regression
and propensity score in high dimension.
}
\author{
Emeline Courtois \cr Maintainer:
Emeline Courtois <emeline.courtois@inserm.fr>
}
