% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt4pv-package.R
\name{data_PV}
\alias{data_PV}
\alias{X}
\alias{Y}
\title{Simulated data for the adapt4pv package}
\format{\describe{
\item{X}{large sparse and binary matrix with 117160 rows and 300 columns.
Drug matrix exposure: each row corresponds to an individual and
each column corresponds to a drug.}
\item{Y}{large spase and binary vector of length 117160. Indicator of
the presence/absence of an adverse event for ech individual.
Only the first 30 drugs (out of the 300)
are associated with the outcome.}
}}
\description{
Simple simulated data, used to demonstrate the features of
functions from adapt4cv package.
}
\examples{

data(ExamplePvData)


}
\keyword{datasets}
